refactor(utility): separate common helper functions
