+++ rust/src/store/best_ledger_store_impl.rs
@@ -15,7 +15,10 @@ use crate::{
         token::TokenAddress,
         Ledger, TokenLedger,
     },
-    utility::store::{from_be_bytes, ledger::best::*, pk_index_key},
+    utility::store::{
+        common::{from_be_bytes, pk_index_key},
+        ledger::best::*,
+    },
 };
 use log::trace;
 use speedb::{DBIterator, IteratorMode};
