+++ rust/src/store/user_command_store_impl.rs
@@ -16,7 +16,8 @@ use crate::{
     constants::millis_to_iso_date_string,
     ledger::public_key::PublicKey,
     utility::store::{
-        command::user::*, from_be_bytes, pk_key_prefix, pk_txn_sort_key_sort, u32_prefix_key,
+        command::user::*,
+        common::{from_be_bytes, pk_key_prefix, pk_txn_sort_key_sort, u32_prefix_key},
     },
 };
 use anyhow::bail;
