+++ rust/src/utility/store/ledger/best.rs
@@ -1,6 +1,6 @@
 use crate::{
     ledger::{public_key::PublicKey, token::TokenAddress},
-    utility::store::{balance_key_prefix, pk_key_prefix, U64_LEN},
+    utility::store::common::{balance_key_prefix, pk_key_prefix, U64_LEN},
 };
 
 /// Key format for storing best ledger accounts
