+++ rust/src/web/graphql/stakes/mod.rs
@@ -10,7 +10,7 @@ use crate::{
     },
     snark_work::store::SnarkStore,
     store::{username::UsernameStore, IndexerStore},
-    utility::store::U32_LEN,
+    utility::store::common::U32_LEN,
     web::graphql::Timing,
 };
 use async_graphql::{ComplexObject, Context, Enum, InputObject, Object, Result, SimpleObject};
