+++ rust/src/web/graphql/top_snarkers.rs
@@ -4,7 +4,7 @@ use crate::{
     ledger::public_key::PublicKey,
     snark_work::store::SnarkStore,
     store::username::UsernameStore,
-    utility::store::{u32_from_be_bytes, U32_LEN, U64_LEN},
+    utility::store::common::{u32_from_be_bytes, U32_LEN, U64_LEN},
 };
 use async_graphql::{Context, Enum, InputObject, Object, Result, SimpleObject};
 use speedb::Direction;
